% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{IRangeKernels}
\alias{IRangeKernels}
\alias{IRangeKernels-package}
\title{IRangeKernels - Kernel density estimation using IRanges}
\description{
The package extends IRanges to allow fast sliding window density estimation
by implementing a Views kernel multiplication function, viewMuls,
to multiply Views by a vector.
}
\author{
Nathan Sheffield
}
\references{
This has not been published, see
\url{github.com/sheffien}
}

