% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RleViews-utils.R
\name{viewMuls}
\alias{viewMuls}
\title{viewMuls}
\usage{
viewMuls(x, na.rm = FALSE, y)
}
\arguments{
\item{x}{Views object.}

\item{na.rm}{remove NAs?}

\item{y}{kernel vector or weights to multiply by; should be a vector matching the length of x.}
}
\description{
Analogous to the IRanges viewSums, but multiplies by a vector; so you can think
of it as a weighted sum, or a kernel density estimate.
}

